+++ pixyz/losses/autoregressive.py
@@ -1,7 +1,6 @@
 from copy import deepcopy
 
 from .losses import Loss
-from .expectations import Expectation
 from ..utils import get_dict_values
 
 
@@ -25,7 +24,7 @@ class IterativeLoss(Loss):
 
         self.slice_step = slice_step
         if self.slice_step:
-            self.step_loss = Expectation(self.slice_step, self.step_loss)
+            self.step_loss = self.step_loss.expectation(self.slice_step)
 
         if input_var is not None:
             self._input_var = input_var
